hypervisor/openstack: fix usage with V3 API (#5)
* hypervisor/openstack: permit specifying openstack_project_name

KeyStone V3 API needs the project name instead of the tenant's name
(such field doesn't exist anymore). Transparently let users pass both
options. At least one needs to be specified. Later on, Fog handles
switching between the V3 and V2 versions depending on the
openstack_auth_url parameter.

* hypervisor/openstack: make it work with newest fog-openstack

The errors have been renamed to Fog::OpenStack::* instead of
Fog::Compute::OpenStack::*. At the same time bump the requirement of
`fog-openstack` because it will not work with earlier versions.

* Revert "hypervisor/openstack: make it work with newest fog-openstack"

This reverts commit e75b94caa8255b7802cbb3d330255801e9a0ff5f.

* beaker-openstack.gemspec: pin version which works

* beaker: add proper check for args

With v2 only certain parameters are allowed, and with v3 - different
ones. Add checks for those. Refactor the v3 check into a different
variable. Fix the tests so that they would pass with the new checks.

* spec/beaker: add missing test for nil

This field should be nil too so add a test for it.

* beaker/openstack: print out more parameters on err

* beaker/openstack: make creating floating IPs optional

* Move `get_ip` to `get_floating_ip`
* Add another parameter `openstack_floating_ip` (bool) which controls if
we want to create a floating IP address for the VM
* If no floating IP is requested, generate a random 10 character
hostname